Changed the baud rate and uart_write function

+++ uart.h
@@ -7,7 +7,7 @@
 
 void uart2_rxtx_interrupt_init();
 void uart2_rxtx_init(void);
-void uart2_write(int ch);
+void uart2_write(uint8_t ch);
 //=============================================================
 void uart2_tx_init(void); // use only in polling method
 // ============================================================
